requirements-dev.txt
azure-functions>=1.18.0,<2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
requests>=2.31.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
